+++ client/src/features/common/LuxRoutes.tsx
@@ -3,7 +3,6 @@ import { Route, Routes, useLocation } from 'react-router-dom'
 import { Alert } from 'react-bootstrap'
 
 import { useGetItemQuery } from '../../redux/api/ml_api'
-import theme from '../../styles/theme'
 import ErrorPage from '../error/ErrorPage'
 import RoutingComponent from '../results/RoutingComponent'
 import Landing from '../landing/LandingPage'
@@ -13,6 +12,8 @@ import CmsRoutingComponent from '../cms/CmsRoutingComponent'
 import { pushClientPageEvent } from '../../lib/pushClientEvent'
 import { getTargetName } from '../../lib/util/uri'
 import { getRouteNames } from '../../config/routerPages'
+import useResizeableWindow from '../../lib/hooks/useResizeableWindow'
+import theme from '../../styles/theme'
 
 import Footer from './Footer'
 
@@ -35,6 +36,9 @@ const RedirectOldProd: React.FC = () => {
 const LuxRoutes: React.FC = () => {
   const { pathname, search } = useLocation()
   const [prevUrl, setPrevUrl] = useState('')
+  const [showMobileAlert, setShowMobileAlert] = useState<boolean>(
+    window.innerWidth < theme.breakpoints.md,
+  )
   const routes = getRouteNames()
   const isNotAnEntityPage = routes.has(pathname)
 
@@ -69,6 +73,8 @@ const LuxRoutes: React.FC = () => {
     setPrevUrl(currentUrl)
   }, [data, isNotAnEntityPage, isSuccess, pathname, prevUrl, routes, search])
 
+  useResizeableWindow(setShowMobileAlert)
+
   return (
     <React.Fragment>
       <RedirectOldProd />
@@ -78,8 +84,12 @@ const LuxRoutes: React.FC = () => {
         <Route path="/*" element={<Header />} />
       </Routes>
       <div className="container-fluid px-0" id="route-container">
-        {window.innerWidth < theme.breakpoints.md && (
-          <Alert variant="info" className="d-flex justify-content-center">
+        {showMobileAlert && (
+          <Alert
+            dismissible
+            variant="info"
+            className="d-flex justify-content-center mb-0"
+          >
             LUX is optimized for desktop use. Some features are not available on
             mobile devices.
           </Alert>