+++ client/src/features/results/ResultsHeader.tsx
@@ -4,25 +4,38 @@ import sanitizeHtml from 'sanitize-html'
 import { Button, ButtonGroup, Col, Row } from 'react-bootstrap'
 import styled from 'styled-components'
 
+import theme from '../../styles/theme'
+import useResizeableWindow from '../../lib/hooks/useResizeableWindow'
 import StyledResultsHeader from '../../styles/features/results/ResultsHeader'
 import StyledHr from '../../styles/shared/Hr'
 import { pushClientEvent } from '../../lib/pushClientEvent'
 import { OverlayKey } from '../../config/cms'
-import { sortBy, sortDirection } from '../../config/sortingOptions'
 import { getParamPrefix } from '../../lib/util/params'
 import EntityResultsDescription from '../cms/EntityResultsDescription'
 import { ResultsTab } from '../../types/ResultsTab'
 import LuxOverlay from '../common/LuxOverlay'
+import MobileSelectedFacets from '../facets/MobileSelectedFacets'
+import { searchScope } from '../../config/searchTypes'
 
-import SortDropdown from './SortDropdown'
-
-// Sequence number used to create a different URL for a random shuffle query
-let seq = 0
+import Sort from './Sort'
 
 const StyledCol = styled(Col)`
-  @media (min-width: 768px) {
+  margin-bottom: 12px;
+  margin-top: 12px;
+
+  @media (min-width: ${theme.breakpoints.md}px) {
     justify-content: flex-end;
     text-align: right;
+    margin-top: 0px;
+    margin-bottom: 0px;
+  }
+`
+
+const StyledDiv = styled.div`
+  display: none;
+
+  @media (min-width: ${theme.breakpoints.md}px) {
+    display: inline;
   }
 `
 
@@ -39,6 +52,12 @@ const ResultsHeader: React.FC<IResultsHeader> = ({
   overlay,
   toggleView = false,
 }) => {
+  const [isMobile, setIsMobile] = useState<boolean>(
+    window.innerWidth < theme.breakpoints.md,
+  )
+
+  useResizeableWindow(setIsMobile)
+
   const [redirect, setRedirect] = useState<boolean>(false)
 
   useEffect(() => {
@@ -67,13 +86,7 @@ const ResultsHeader: React.FC<IResultsHeader> = ({
   const sortDirectionParamValue = queryString.has(sortName)
     ? (queryString.get(sortName)?.split(':')[1] as string)
     : undefined
-  const sortByOptions = sortBy[tab]
   const descriptiveText = EntityResultsDescription(overlay) || ''
-
-  const [sortBySelection, setSortBySelection] = useState<string>(sort || '')
-  const [selectedSortDirection, setSelectedSortDirection] = useState<string>(
-    sortDirectionParamValue || 'desc',
-  )
 
   // toggle view between list and image view
   const changeView = (selectedView: string): void => {
@@ -85,42 +98,9 @@ const ResultsHeader: React.FC<IResultsHeader> = ({
 
     queryString.set('view', selectedView)
     navigate({
-      pathname: `/view/results/${tab !== undefined ? tab : 'objects'}`,
+      pathname,
       search: `?${queryString.toString()}`,
     })
-  }
-
-  const handleSortDirectionSelection = (value: string): void => {
-    pushClientEvent(
-      'Results Sort',
-      'Selected',
-      `Sort By ${sortBySelection}:${value}`,
-    )
-    setSelectedSortDirection(value)
-    if (sort !== undefined) {
-      setSortBySelection(sort)
-      // set query string params
-      queryString.set(sortName, `${sortBySelection}:${value}`)
-      const searchQ = queryString.toString()
-      navigate(`${pathname}?${searchQ}`)
-    }
-  }
-
-  const handleSortSelection = (value: string): void => {
-    pushClientEvent(
-      'Results Sort',
-      'Selected',
-      `Sort By ${value}:${selectedSortDirection}`,
-    )
-    queryString.delete('rnd')
-    if (value === 'random') {
-      queryString.set('rnd', String(seq))
-      seq += 1
-    }
-    queryString.set(sortName, `${value}:${selectedSortDirection}`)
-    const searchQ = queryString.toString()
-    setSortBySelection(value)
-    navigate(`${pathname}?${searchQ}`)
   }
 
   return (
@@ -131,12 +111,14 @@ const ResultsHeader: React.FC<IResultsHeader> = ({
             className="mb-0"
             data-testid="results-header-title"
           >
-            {total} {label} results
+            <StyledDiv>
+              {total} {label} results
+            </StyledDiv>
             {(tab === 'objects' || tab === 'works') && <LuxOverlay />}
           </StyledResultsHeader>
         </Col>
       </Row>
-      <Row>
+      <Row className="px-2">
         <Col xs={12} sm={12} md={6} lg={6} xl={6}>
           <div
             className="descriptiveText"
@@ -152,7 +134,7 @@ const ResultsHeader: React.FC<IResultsHeader> = ({
           md={6}
           lg={6}
           xl={6}
-          className="d-flex align-items-center"
+          className="d-flex align-items-end"
           data-testid="results-header-options"
         >
           <ButtonGroup>
@@ -161,10 +143,16 @@ const ResultsHeader: React.FC<IResultsHeader> = ({
                 {toggleView && (
                   <Button
                     type="button"
-                    className="btn btn-light mx-2 text-center"
+                    className="btn text-center h-100"
                     onClick={() =>
                       changeView(currentView === 'list' ? 'grid' : 'list')
                     }
+                    style={{
+                      borderRadius: theme.border.radius,
+                      backgroundColor: theme.color.lightGray,
+                      color: theme.color.trueBlack,
+                      border: theme.color.trueBlack,
+                    }}
                     data-testid={
                       currentView === 'list'
                         ? 'switch-to-grid-view-button'
@@ -185,31 +173,19 @@ const ResultsHeader: React.FC<IResultsHeader> = ({
                   </Button>
                 )}
               </div>
-              <div className="flex-grow-1 ms-3">
-                <SortDropdown
-                  options={sortByOptions}
-                  handleChange={handleSortSelection}
-                  className="sortingDropdown"
-                  id="sorting-dropdown"
-                  selected={sort}
-                  label="Sort By"
-                  headerText="Sort By"
-                />
-              </div>
-              <div className="flex-grow-1 ms-3">
-                <SortDropdown
-                  options={sortDirection}
-                  handleChange={handleSortDirectionSelection}
-                  className="sortAscOrDesc"
-                  id="sort-asc-or-desc"
-                  selected={sortDirectionParamValue}
-                  label={selectedSortDirection}
-                  headerText="Sorting Direction"
-                />
-              </div>
+              <Sort />
             </div>
           </ButtonGroup>
         </StyledCol>
+        {isMobile && (
+          <MobileSelectedFacets
+            tab={tab}
+            scope={searchScope[tab]}
+            search={search}
+            selectedSortBy={sort}
+            selectedSortDirection={sortDirectionParamValue}
+          />
+        )}
       </Row>
       <StyledHr width="100%" className="my-2" />
     </React.Fragment>