+++ client/src/styles/global.ts
@@ -27,32 +27,60 @@ const GlobalStyle = createGlobalStyle`
     font-size: 3.1em;
     color: #222222;
     letter-spacing: -2px;
-    line-height: 72px;
-    font-weight: ${theme.font.weight.light};
+    font-size: ${theme.font.mobile.h1.size};
+    line-height: ${theme.font.mobile.h1.lineHeight};
+    font-weight: ${theme.font.mobile.h1.weight};
+
+    @media (min-width: ${theme.breakpoints.md}px) {
+      font-size: ${theme.font.desktop.h1.size};
+      line-height: ${theme.font.desktop.h1.lineHeight};
+      font-weight: ${theme.font.desktop.h1.weight};
+    }
   }
 
   h2 {
     color: #222222;
     letter-spacing: 0;
     text-align: left;
-    line-height: 56px;
-    font-weight: 200;
+    font-size: ${theme.font.mobile.h2.size};
+    line-height: ${theme.font.mobile.h2.lineHeight};
+    font-weight: ${theme.font.mobile.h2.weight};
+
+    @media (min-width: ${theme.breakpoints.md}px) {
+      font-size: ${theme.font.desktop.h2.size};
+      line-height: ${theme.font.desktop.h2.lineHeight};
+      font-weight: ${theme.font.desktop.h2.weight};
+    }
   }
 
   h3 {
     color: #222222;
     letter-spacing: 0;
     text-align: left;
-    line-height: 40px;
-    font-weight: 200;
+    font-size: ${theme.font.mobile.h3.size};
+    line-height: ${theme.font.mobile.h3.lineHeight};
+    font-weight: ${theme.font.mobile.h3.weight};
+
+    @media (min-width: ${theme.breakpoints.md}px) {
+      font-size: ${theme.font.desktop.h3.size};
+      line-height: ${theme.font.desktop.h3.lineHeight};
+      font-weight: ${theme.font.desktop.h3.weight};
+    }
   }
 
   h4 {
     color: #222222;
     letter-spacing: 0;
     text-align: left;
-    line-height: 32px;
-    font-weight: 700;
+    font-size: ${theme.font.mobile.h4.size};
+    line-height: ${theme.font.mobile.h4.lineHeight};
+    font-weight: ${theme.font.mobile.h4.weight};
+
+    @media (min-width: ${theme.breakpoints.md}px) {
+      font-size: ${theme.font.desktop.h4.size};
+      line-height: ${theme.font.desktop.h4.lineHeight};
+      font-weight: ${theme.font.desktop.h4.weight};
+    }
   }
 
   h5 {
@@ -60,15 +88,30 @@ const GlobalStyle = createGlobalStyle`
     letter-spacing: 0;
     text-align: left;
     line-height: 24px;
-    font-weight: 700;
+    font-size: ${theme.font.mobile.h5.size};
+    line-height: ${theme.font.mobile.h5.lineHeight};
+    font-weight: ${theme.font.mobile.h5.weight};
+
+    @media (min-width: ${theme.breakpoints.md}px) {
+      font-size: ${theme.font.desktop.h5.size};
+      line-height: ${theme.font.desktop.h5.lineHeight};
+      font-weight: ${theme.font.desktop.h5.weight};
+    }
   }
 
   p {
-    color: #222222;
+    color: ${theme.color.black};
     letter-spacing: 0;
     text-align: left;
-    line-height: 24px;
-    font-weight: 200;
+    font-size: ${theme.font.mobile.bodyLight.size};
+    line-height: ${theme.font.mobile.bodyLight.lineHeight};
+    font-weight: ${theme.font.mobile.bodyLight.weight};
+
+    @media (min-width: ${theme.breakpoints.md}px) {
+      font-size: ${theme.font.desktop.bodyLight.size};
+      line-height: ${theme.font.desktop.bodyLight.lineHeight};
+      font-weight: ${theme.font.desktop.bodyLight.weight};
+    }
   }
 
   ul {