+++ client/src/styles/shared/Hr.ts
@@ -6,6 +6,7 @@ interface IProps {
   borderWidth?: string
   color?: string
   width?: string
+  hiddenOnDesktop?: boolean
 }
 
 const Hr = styled.hr<IProps>`
@@ -15,6 +16,17 @@ const Hr = styled.hr<IProps>`
   padding: 0px 5px;
   width: ${(props) => (props.width ? props.width : '98%')};
   margin: auto;
+
+  &.footerBlocks {
+    margin-bottom: 1rem;
+    @media (min-width: ${theme.breakpoints.lg}px) {
+      display: none;
+    }
+  }
+
+  @media (min-width: ${theme.breakpoints.md}px) {
+    display: ${(props) => (props.hiddenOnDesktop ? 'none' : 'block')};
+  }
 `
 
 export default Hr